Adds kpt live install-resource-group command
